+++ FerroVelho/Consulta/fm_notasCompra.cs
@@ -26,6 +26,11 @@ namespace FerroVelho
             consulta();
         }
 
+        private bool IsPostgresMode
+        {
+            get { return DataContextFactory.IsPostgresConnectionString(DataContextFactory.conexaoImp); }
+        }
+
         private void bt_pesquisa_Click(object sender, EventArgs e)
         {
             consulta();
@@ -35,10 +40,19 @@ namespace FerroVelho
         {
             DateTime dataI = dt_inicio.Value;
             DateTime dataF = dt_fim.Value;
-            
-            string comando = "SELECT * FROM tb_compra " +
-                "WHERE tb_compra.data_compra between '" + dataI.Date + "' AND '" + dataF.Date.Add(new TimeSpan(23,59,59)) + "'";
-            DataTable dt = DataContextFactory.Filtrar(comando);
+
+            DataTable dt;
+            if (IsPostgresMode)
+            {
+                dt = DataContextFactory.ListarComprasPostgres(dataI.Date, dataF.Date.Add(new TimeSpan(23, 59, 59)), null);
+            }
+            else
+            {
+                string comando = "SELECT * FROM tb_compra " +
+                    "WHERE tb_compra.data_compra between '" + dataI.Date + "' AND '" + dataF.Date.Add(new TimeSpan(23, 59, 59)) + "'";
+                dt = DataContextFactory.Filtrar(comando);
+            }
+
             tb_vendaDataGridView.DataSource = dt;
             tb_vendaDataGridView.DataMember = dt.TableName;
             if (tb_vendaDataGridView.Rows.Count >= 1)
@@ -53,7 +67,14 @@ namespace FerroVelho
             try
             {
                 int id = Convert.ToInt32(tb_vendaDataGridView.CurrentRow.Cells[0].Value);
-                this.tb_itemcBindingSource.DataSource = DataContextFactory.DataContext.tb_itemc.Where(x => x.id_compra == id);
+                if (IsPostgresMode)
+                {
+                    this.tb_itemcBindingSource.DataSource = DataContextFactory.ListarItensCompraPostgres(id);
+                }
+                else
+                {
+                    this.tb_itemcBindingSource.DataSource = DataContextFactory.DataContext.tb_itemc.Where(x => x.id_compra == id);
+                }
             }
             catch
             {
@@ -65,9 +86,18 @@ namespace FerroVelho
         {            
             try
             {
-                string comando = "SELECT * FROM tb_compra " +
-                "WHERE tb_compra.id_compra = " + Convert.ToInt32(txt_notaFiscal.Text);
-                DataTable dt = DataContextFactory.Filtrar(comando);
+                DataTable dt;
+                if (IsPostgresMode)
+                {
+                    dt = DataContextFactory.ListarComprasPostgres(null, null, Convert.ToInt32(txt_notaFiscal.Text));
+                }
+                else
+                {
+                    string comando = "SELECT * FROM tb_compra " +
+                    "WHERE tb_compra.id_compra = " + Convert.ToInt32(txt_notaFiscal.Text);
+                    dt = DataContextFactory.Filtrar(comando);
+                }
+
                 tb_vendaDataGridView.DataSource = dt;
                 tb_vendaDataGridView.DataMember = dt.TableName;
 
@@ -81,8 +111,17 @@ namespace FerroVelho
             }
             catch
             {
-                string comando = "SELECT * FROM tb_compra ";
-                DataTable dt = DataContextFactory.Filtrar(comando);
+                DataTable dt;
+                if (IsPostgresMode)
+                {
+                    dt = DataContextFactory.ListarComprasPostgres(null, null, null);
+                }
+                else
+                {
+                    string comando = "SELECT * FROM tb_compra ";
+                    dt = DataContextFactory.Filtrar(comando);
+                }
+
                 tb_vendaDataGridView.DataSource = dt;
                 tb_vendaDataGridView.DataMember = dt.TableName;
                 if (tb_vendaDataGridView.Rows.Count >= 1)
@@ -99,11 +138,18 @@ namespace FerroVelho
             try
             {
                 int id = Convert.ToInt32(tb_vendaDataGridView.CurrentRow.Cells[0].Value);
-                comando = new SqlCommand();
-                comando.CommandType = CommandType.Text;
-                comando.CommandText = "Delete from tb_compra WHERE id_compra=@id";
-                comando.Parameters.AddWithValue("@id", id);
-                DataContextFactory.CRUD(comando);
+                if (IsPostgresMode)
+                {
+                    DataContextFactory.ExcluirCompraPostgres(id);
+                }
+                else
+                {
+                    comando = new SqlCommand();
+                    comando.CommandType = CommandType.Text;
+                    comando.CommandText = "Delete from tb_compra WHERE id_compra=@id";
+                    comando.Parameters.AddWithValue("@id", id);
+                    DataContextFactory.CRUD(comando);
+                }
 
                 consulta();
                 MessageBox.Show("Excluido com sucesso!");
@@ -165,6 +211,11 @@ namespace FerroVelho
 
         private string usuarioDAO(int id)
         {
+            if (IsPostgresMode)
+            {
+                return DataContextFactory.BuscarNomeUsuarioPostgres(id);
+            }
+
             tb_usuario usuario = new tb_usuario();            
 
             comando = new SqlCommand();
@@ -189,6 +240,11 @@ namespace FerroVelho
 
         private string clienteDAO(int id)
         {
+            if (IsPostgresMode)
+            {
+                return DataContextFactory.BuscarNomeClientePostgres(id);
+            }
+
             tb_cliente cliente = new tb_cliente();
 
             comando = new SqlCommand();
